# backend/src/main/resources/api-docs.yaml
openapi: 3.0.3
info:
  title: MyTrip API
  description: API documentation for MyTrip travel booking platform
  version: 1.0.0
  contact:
    name: MyTrip Support
    email: support@mytrip.com
servers:
  - url: http://localhost:8080/api
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /hotels:
    get:
      summary: Get all hotels
      responses:
        '200':
          description: List of hotels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Hotel'
    post:
      summary: Create a new hotel
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Hotel'
      responses:
        '201':
          description: Hotel created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hotel'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (admin required)
  /hotels/{id}:
    get:
      summary: Get hotel by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Hotel details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hotel'
        '404':
          description: Hotel not found
    put:
      summary: Update hotel
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Hotel'
      responses:
        '200':
          description: Hotel updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hotel'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (admin required)
        '404':
          description: Hotel not found
    delete:
      summary: Delete hotel
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Hotel deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (admin required)
        '404':
          description: Hotel not found
...
  /rooms/hotel/{hotelId}:
    get:
      summary: Get rooms by hotel ID
      parameters:
        - in: path
          name: hotelId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of rooms for the hotel
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HotelRoom'
        '404':
          description: Hotel not found
  /flights:
    get:
      summary: Get all flights
      responses:
        '200':
          description: List of flights
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Flight'
    post:
      summary: Create a new flight
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Flight'
      responses:
        '201':
          description: Flight created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flight'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (admin required)
  /flights/{id}:
    get:
      summary: Get flight by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Flight details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flight'
        '404':
          description: Flight not found
    put:
      summary: Update flight
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Flight'
      responses:
        '200':
          description: Flight updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flight'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (admin required)
        '404':
          description: Flight not found
    delete:
      summary: Delete flight
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Flight deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (admin required)
        '404':
          description: Flight not found
  /flights/{id}/seats:
    get:
      summary: Get seats for a flight
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of seats for the flight
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FlightSeat'
        '404':
          description: Flight not found
  /flights/{id}/status:
    get:
      summary: Get live flight status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Flight status
          content:
            application/json:
              schema:
                type: object
                properties:
                  flightId:
                    type: integer
                    format: int64
                  status:
                    type: string
                    enum: [ON_TIME, DELAYED, BOARDING, DEPARTED, ARRIVED, CANCELLED]
                  delayReason:
                    type: string
                    nullable: true
                  lastUpdated:
                    type: string
                    format: date-time
        '404':
          description: Flight not found
  /seats:
    get:
      summary: Get all flight seats
      responses:
        '200':
          description: List of flight seats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FlightSeat'
    post:
      summary: Create a new flight seat
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlightSeat'
      responses:
        '201':
          description: Flight seat created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlightSeat'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (admin required)
  /seats/{id}:
    get:
      summary: Get flight seat by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Flight seat details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlightSeat'
        '404':
          description: Flight seat not found
    put:
      summary: Update flight seat
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlightSeat'
      responses:
        '200':
          description: Flight seat updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlightSeat'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (admin required)
        '404':
          description: Flight seat not found
    delete:
      summary: Delete flight seat
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Flight seat deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (admin required)
        '404':
          description: Flight seat not found
  /seats/flight/{flightId}:
    get:
      summary: Get seats by flight ID
      parameters:
        - in: path
          name: flightId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of seats for the flight
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FlightSeat'
        '404':
          description: Flight not found
...
  /reviews/hotel/{hotelId}:
    get:
      summary: Get reviews for a hotel
      parameters:
        - in: path
          name: hotelId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of reviews for the hotel
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
        '404':
          description: Hotel not found
  /reviews/flight/{flightId}:
    get:
      summary: Get reviews for a flight
      parameters:
        - in: path
          name: flightId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of reviews for the flight
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
        '404':
          description: Flight not found
...
  /price-history/hotel/{hotelId}:
    get:
      summary: Get price history for a hotel
      parameters:
        - in: path
          name: hotelId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Price history for the hotel
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PriceHistory'
        '404':
          description: Hotel not found
  /price-history/flight/{flightId}:
    get:
      summary: Get price history for a flight
      parameters:
        - in: path
          name: flightId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Price history for the flight
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PriceHistory'
        '404':
          description: Flight not found
...